+++ changelog.md
@@ -2,7 +2,8 @@
 
 | Version | Date | Description |
 | --- | --- | --- |
-| 0.1.00002 | 18/11/2024 | ACL Rules adn URL Filterings are removed at integration reload if not available anymore |
+| 0.1.00003 | 18/11/2024 | ACL Rules and URL Filterings sensors are displayed again, removed Disabled switch again to prevent confusion |
+| 0.1.00002 | 18/11/2024 | ACL Rules and URL Filterings are removed at integration reload if not available anymore |
 | 0.1.00001 | 18/11/2024 | added DEvice Tracker for Omada Devices and Omada Clients |
 | 0.0.5 | 17/11/2024 | fixing issues with creating / updating Omada Devices and Omada Clients |
 | 0.0.4 | 05/11/2024 | acl / url filtering switch names |